pyproject.toml
[project]
name = "cue-designer"
version = "0.1.0"
description = "Django REST application for designing pool cue butt sections"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "shapely>=2.0.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cue_designer.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "--reuse-db --nomigrations --cov=apps --cov-report=html"

[tool.black]
line-length = 88
target-version = ['py314']

[tool.ruff]
line-length = 88
target-version = "py314"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.14"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true